Chore: separate channel contorller and router
채널의 라우터와 컨트롤러 파일을 분리했습니다.
